+++ exercises/exercise_4/question_5/src/houghElliptical.cpp
@@ -51,7 +51,7 @@ void HoughElliptical( int, void* )
 void* executeHoughElliptical( void* args )
 {
    int dev              = 0;
-   uint16_t frame_count = 0;
+   uint32_t frame_count = 0;
    //float frame_rate;
    //struct timespec start_time, stop_time, diff_time;
